Build-test changes

+++ backend/webapi/Infrastructure/HttpClients/Jum/IJumClient.cs
@@ -6,7 +6,7 @@ public interface IJumClient
 {
     Task<JustinUser?> GetJumUserAsync(string username);
     Task<Participant?> GetJumUserAsync(string username, string accessToken);
-    Task<JustinUser?> GetJumUserByPartIdAsync(long partId);
+    Task<JustinUser?> GetJumUserByPartIdAsync(decimal partId);
     Task<Participant?> GetJumUserByPartIdAsync(decimal partId, string accessToken);
     Task<bool> IsJumUser(Participant? justinUser, Party party);
 }
